strand-cam: default saving to MP4
Support for saving video .mp4 files in Strand Camera, `braid-process-video`,
`strand-convert`, and other utilities. Video is encoded with the H.264 codec and
metadata, including precise timestamps, are stored in the h264 stream.
Uncompressed 8bit and high dynamic range images can be saved using
[less-avc](https://crates.io/crates/less-avc). With [appropriate NVENC
hardware](https://developer.nvidia.com/video-encode-and-decode-gpu-support-matrix-new),
hardware-accelerated encoding is also supported.

+++ fmf/src/lib.rs
@@ -192,7 +192,7 @@ impl<F: Write + Seek> FMFWriterInner<F> {
         let bytes_per_pixel = pixel_format.bits_per_pixel() / 8;
 
         let row_bytes = w as usize * bytes_per_pixel as usize;
-        let chunksize = (row_bytes * h as usize + 8) as usize;
+        let chunksize = row_bytes * h as usize + 8;
 
         let mut pos = 0;
         f.write_u32::<LittleEndian>(3)?;
